+++ pinery-miso/src/main/resources/ca/on/oicr/pinery/lims/miso/queryAllAssays.sql
@@ -2,6 +2,8 @@ SELECT assayId
   ,alias AS name
   ,description
   ,version
+  ,draft
+  ,archived
   ,caseTargetDays
   ,receiptTargetDays
   ,extractionTargetDays