+++ sqlstore/src/main/resources/db/migration/V1760__auto_main.sql
@@ -1,0 +1,3 @@
+-- assay_draft
+ALTER TABLE Assay ADD COLUMN draft BOOLEAN NOT NULL DEFAULT FALSE;
+